+++ History.md
@@ -1,4 +1,14 @@
 
+0.0.5 / 2013-11-18
+==================
+
+ * add gruntfile, jshintrc and pacakge.json
+ * add ns alias to prefix
+ * add del alias to unset
+ * add reset alias to clear
+ * set now returns 0 when it fails and data.length otherwise
+ * Update Readme.md
+
 0.0.4 / 2013-07-25
 ==================